% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg.R
\name{fg_context}
\alias{fg_context}
\title{Generate LLM-friendly context for the fangraphs.com package}
\usage{
fg_context()
}
\value{
Character string (invisibly), also printed
}
\description{
Generate LLM-friendly context for the fangraphs.com package
}
